requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
openai-whisper>=20231117
sendgrid>=6.9.0
jinja2>=3.0.0
pydantic>=1.8.0
python-multipart==0.0.9
aiosmtplib==3.0.1
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.1.2
 